Network: SQLite supports only floats :/
Goodbye precision

+++ server/src/database.cc
@@ -181,11 +181,11 @@ network_from_row(void* optional_network, int rows,
         if (std::strcmp(*column_name, "name") == 0)
           network.name = *column_text;
         else if (std::strcmp(*column_name, "lat") == 0)
-          network.lat = boost::lexical_cast<double>(*column_text);
+          network.lat = boost::lexical_cast<float>(*column_text);
         else if (std::strcmp(*column_name, "lon") == 0)
-          network.lon = boost::lexical_cast<double>(*column_text);
+          network.lon = boost::lexical_cast<float>(*column_text);
         else if (std::strcmp(*column_name, "strength") == 0)
-          network.strength = boost::lexical_cast<double>(*column_text);
+          network.strength = boost::lexical_cast<float>(*column_text);
         else
           throw std::runtime_error{
               std::string{"Unexpected column: "} + *column_name};
